refactor network validator
