+++ client/src/androidUnitTest/kotlin/org/apptank/horus/client/TestCase.kt
@@ -11,7 +11,7 @@ import org.apptank.horus.client.extensions.createSQLInsert
 import org.apptank.horus.client.extensions.execute
 import org.apptank.horus.client.extensions.prepareSQLValueAsString
 import org.apptank.horus.client.di.IDatabaseDriverFactory
-import org.apptank.horus.client.di.INetworkValidator
+import org.apptank.horus.client.connectivity.INetworkValidator
 import org.apptank.horus.client.migration.network.service.IMigrationService
 import org.apptank.horus.client.sync.network.service.ISynchronizationService
 import org.apptank.horus.client.tasks.RetrieveDatabaseSchemeTask
