+++ client/src/androidUnitTest/kotlin/org/apptank/horus/client/sync/manager/RemoteSynchronizatorManagerTest.kt
@@ -8,7 +8,7 @@ import org.apptank.horus.client.control.helper.ISyncControlDatabaseHelper
 import org.apptank.horus.client.control.SyncControl
 import org.apptank.horus.client.eventbus.EventBus
 import org.apptank.horus.client.eventbus.EventType
-import org.apptank.horus.client.di.INetworkValidator
+import org.apptank.horus.client.connectivity.INetworkValidator
 import org.apptank.horus.client.sync.network.service.ISynchronizationService
 import io.mockative.Mock
 import io.mockative.any
