+++ client/src/androidUnitTest/kotlin/org/apptank/horus/client/sync/manager/SyncFileUploadedManagerTest.kt
@@ -9,7 +9,7 @@ import io.mockative.mock
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.delay
 import kotlinx.coroutines.runBlocking
-import org.apptank.horus.client.di.INetworkValidator
+import org.apptank.horus.client.connectivity.INetworkValidator
 import org.apptank.horus.client.eventbus.EventBus
 import org.apptank.horus.client.eventbus.EventType
 import org.apptank.horus.client.sync.upload.data.SyncFileResult
