+++ client/src/androidUnitTest/kotlin/org/apptank/horus/client/tasks/RetrieveDataSharedTaskTest.kt
@@ -15,7 +15,7 @@ import kotlinx.datetime.Clock
 import org.apptank.horus.client.TestCase
 import org.apptank.horus.client.base.DataResult
 import org.apptank.horus.client.control.helper.IDataSharedDatabaseHelper
-import org.apptank.horus.client.di.INetworkValidator
+import org.apptank.horus.client.connectivity.INetworkValidator
 import org.apptank.horus.client.sync.network.dto.SyncDTO
 import org.apptank.horus.client.sync.network.service.ISynchronizationService
 import org.junit.Assert
