+++ client/src/androidUnitTest/kotlin/org/apptank/horus/client/tasks/SynchronizeDataTaskTest.kt
@@ -5,7 +5,7 @@ import org.apptank.horus.client.auth.HorusAuthentication
 import org.apptank.horus.client.base.DataResult
 import org.apptank.horus.client.control.helper.ISyncControlDatabaseHelper
 import org.apptank.horus.client.control.helper.IOperationDatabaseHelper
-import org.apptank.horus.client.di.INetworkValidator
+import org.apptank.horus.client.connectivity.INetworkValidator
 import org.apptank.horus.client.sync.network.dto.SyncDTO
 import org.apptank.horus.client.sync.network.service.ISynchronizationService
 import io.mockative.Mock
