+++ client/src/commonMain/kotlin/org/apptank/horus/client/sync/manager/SyncFileUploadedManager.kt
@@ -9,7 +9,7 @@ import kotlinx.coroutines.launch
 import org.apptank.horus.client.base.Callback
 import org.apptank.horus.client.base.DataResult
 import org.apptank.horus.client.base.SuspendedCallback
-import org.apptank.horus.client.di.INetworkValidator
+import org.apptank.horus.client.connectivity.INetworkValidator
 import org.apptank.horus.client.eventbus.EventBus
 import org.apptank.horus.client.eventbus.EventType
 import org.apptank.horus.client.extensions.info
