+++ client/src/commonMain/kotlin/org/apptank/horus/client/tasks/SynchronizeInitialDataTask.kt
@@ -17,7 +17,7 @@ import org.apptank.horus.client.control.SyncControl
 import org.apptank.horus.client.data.Horus
 import org.apptank.horus.client.control.helper.IOperationDatabaseHelper
 import org.apptank.horus.client.database.struct.toRecordsInsert
-import org.apptank.horus.client.di.INetworkValidator
+import org.apptank.horus.client.connectivity.INetworkValidator
 import org.apptank.horus.client.eventbus.EventBus
 import org.apptank.horus.client.eventbus.EventType
 import org.apptank.horus.client.extensions.info
